@@ -1,9 +1,17 @@
-
 /**
  * Box Circle.
  * @class Circle
  */
-class Circle {
+import {Point} from "./Point";
+import {Box} from "./Box";
+import {Shape, UserCustomData} from "./types";
+
+export class Circle  implements Shape{
+    readonly x: number;
+    readonly y: number;
+    readonly r: number;
+    readonly rPow2: number;
+    readonly data: UserCustomData;
 
     /**
      * Circle constructor;
@@ -13,15 +21,15 @@ class Circle {
      * @param {number} r - Radius of the circle.
      * @param {*} [data] - Data to store along the circle.
      */
-    constructor(x, y, r, data) {
+    constructor(x: number, y: number, r: number, data?: UserCustomData) {
         this.x = x;
         this.y = y;
         this.r = r;
         this.rPow2 = this.r * this.r; // To avoid square roots
-        if (data) this.data = data;
+        this.data = data;
     }
 
-    _euclideanDistancePow2(point1, point2) {
+    private euclideanDistancePow2(point1: Point, point2: Point): number {
         return Math.pow((point1.x - point2.x), 2) + Math.pow((point1.y - point2.y), 2);
     }
 
@@ -30,8 +38,8 @@ class Circle {
      * @param {Point|Object} point - The point to test if it is contained in the circle.
      * @returns {boolean} - True if the point is contained in the circle, otherwise false.
      */
-    contains(point) {
-        return this._euclideanDistancePow2(point, this) <= this.rPow2;
+    contains(point: Point): boolean {
+        return this.euclideanDistancePow2(point, this) <= this.rPow2;
     }
 
     /**
@@ -39,14 +47,12 @@ class Circle {
      * @param {Box|Object} range - The box to test the intersection with.
      * @returns {boolean} - True if it intersects, otherwise false.
      */
-    intersects(range) {
-        const Max = (a, b) => a >= b ? a : b;
-        const Min = (a, b) => a <= b ? a : b;
+    intersects(range: Box): boolean {
+        const Max = (a: number, b: number): number => a >= b ? a : b;
+        const Min = (a: number, b: number): number => a <= b ? a : b;
 
         const dX = this.x - Max(range.x, Min(this.x, range.x + range.w));
         const dY = this.y - Max(range.y, Min(this.y, range.y + range.h));
         return (dX * dX + dY * dY) <= (this.rPow2);
     }
 }
-
-module.exports = Circle;
