@@ -1,9 +1,13 @@
-
 /**
  * Point class.
  * @class Point
  */
-class Point {
+import {UserCustomData} from "./types";
+
+export class Point {
+    readonly x: number;
+    readonly y: number;
+    readonly data: UserCustomData;
 
     /**
      * Point constructor.
@@ -12,12 +16,10 @@ class Point {
      * @param {number} y - Y coordinate of the point.
      * @param {*} [data] - Data to store along the point.
      */
-    constructor(x, y, data) {
+    constructor(x: number, y: number, data?: UserCustomData) {
         this.x = x;
         this.y = y;
-        if (data) this.data = data;
+        this.data = data;
     }
 
 }
-
-module.exports = Point;
